+++ bench/go/cmd/bench/go.mod
@@ -1,6 +1,6 @@
 module github.com/microsoft/dev-tunnels-ssh/bench/go/cmd/bench
 
-go 1.17
+go 1.21
 
 require github.com/microsoft/dev-tunnels-ssh/src/go v0.0.0
