+++ test/README.md
@@ -11,12 +11,12 @@ test/
 │   └── Ssh.Test/           # 16 test files, ~185 tests
 ├── ts/                     # TypeScript unit tests (Mocha + @testdeck)
 │   └── ssh-test/           # 18 test files, ~178 tests
-├── go/                     # Go integration tests + interop helpers
-│   ├── ssh-test/           # 16 integration test files, ~158 tests
-│   └── interop/            # E2E helper binaries (Go, TS, C#)
-│       ├── go/             # Go interop helper (all modes)
-│       ├── ts/             # TS interop helper (echo only)
-│       └── cs/             # C# interop helper (echo only)
+├── go/                     # Go integration tests
+│   └── ssh-test/           # 16 integration test files, ~158 tests
+├── interop/                # E2E helper binaries (Go, TS, C#)
+│   ├── go/                 # Go interop helper (all modes)
+│   ├── ts/                 # TS interop helper (echo only)
+│   └── cs/                 # C# interop helper (echo only)
 ├── e2e/                    # E2E shell scripts (~48 tests)
 │   └── README.md           # Detailed E2E documentation
 ├── data/                   # Shared test keys (RSA, ECDSA, various formats)
@@ -192,7 +192,7 @@ Shared key material used across all platforms:
 Generated by `New-TestKeys.ps1`. Not all platforms support all formats — C#
 supports SSH2/OpenSSH formats, TS supports JWK, Go supports PKCS#8/SEC1.
 
-### Interop Helpers (`test/go/interop/`)
+### Interop Helpers (`test/interop/`)
 
 Small CLI programs wrapping each implementation for E2E tests:
