--- test/go/interop/go/go.mod
@@ -1,7 +1,0 @@
-module github.com/microsoft/dev-tunnels-ssh/test/go/interop/go
-
-go 1.17
-
-require github.com/microsoft/dev-tunnels-ssh/src/go v0.0.0
-
-replace github.com/microsoft/dev-tunnels-ssh/src/go => ../../../../src/go
