+++ test/interop/cs/InteropHelper.csproj
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<OutputType>Exe</OutputType>
+		<TargetFramework>net8.0</TargetFramework>
+		<AssemblyName>InteropHelper</AssemblyName>
+		<RootNamespace>InteropHelper</RootNamespace>
+		<IsPackable>false</IsPackable>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="$(SshSourceDirectory)Ssh/Ssh.csproj" />
+		<ProjectReference Include="$(SshSourceDirectory)Ssh.Keys/Ssh.Keys.csproj" />
+		<ProjectReference Include="$(SshSourceDirectory)Ssh.Tcp/Ssh.Tcp.csproj" />
+	</ItemGroup>
+
+</Project>
