+++ test/interop/go/go.mod
@@ -1,0 +1,7 @@
+module github.com/microsoft/dev-tunnels-ssh/test/interop/go
+
+go 1.21
+
+require github.com/microsoft/dev-tunnels-ssh/src/go v0.0.0
+
+replace github.com/microsoft/dev-tunnels-ssh/src/go => ../../../src/go
